+++ src/app/Playground/NavigationBarColorSample.xaml.cs
@@ -5,6 +5,7 @@ public partial class NavigationBarColorSample
     public NavigationBarColorSample()
     {
         InitializeComponent();
+        SetValue(NavigationPage.BarBackgroundColorProperty, Microsoft.Maui.Graphics.Colors.Black);
     }
 
     private async void OnCloseClicked(object sender, EventArgs e)