+++ src/library/DIPS.Mobile.UI/Components/Pages/Android/ContentPage.toolbar.cs
@@ -101,6 +101,10 @@ public partial class ContentPage
         BottomToolbar?.Handler?.DisconnectHandler();
     }
 
+    private partial void RefreshStatusBarTextOnPlatform(NavigationPage navigationPage)
+    {
+    }
+
     private static ViewGroup? FindFrameLayoutParent(AView? view)
     {
         var current = view?.Parent;